+++ packages/backend/src/config/lightdashConfig.mock.ts
@@ -292,4 +292,7 @@ export const lightdashConfigMock: LightdashConfig = {
     partialCompilation: {
         enabled: false,
     },
+    funnelBuilder: {
+        enabled: false,
+    },
 };